requirements.txt
streamlit
langchain
langchain_core
langchain_community
langchain_openai
beautifulsoup4
python-dotenv
chromadb
pytest
neo4j
openai
py2neo
spacy
newspaper3k
asyncio
rake-nltk
pytest-cov
readability-lxml
tiktoken 
langchain-cohere 
langchainhub 
langchain 
langgraph  
tavily-python
pytest-asyncio
psutil
matplotlib
pandas
plotly
requests
fastapi
uvicorn
pydantic-settings
jinja2

pycountry
python-us
# python -m spacy download en_core_web_sm